+++ app/grandchallenge/challenges/views.py
@@ -31,7 +31,6 @@ from grandchallenge.core.guardian import (
     PermissionListMixin,
 )
 from grandchallenge.datatables.views import Column, PaginatedTableListView
-from grandchallenge.invoices.models import Invoice
 from grandchallenge.publications.models import Publication
 from grandchallenge.subdomains.mixins import ChallengeSubdomainObjectMixin
 from grandchallenge.subdomains.utils import reverse, reverse_lazy
@@ -357,36 +356,3 @@ class OnboardingTaskComplete(
             msg += "as incomplete"
 
         return msg
-
-
-class InvoiceList(
-    LoginRequiredMixin,
-    ObjectPermissionRequiredMixin,
-    ListView,
-):
-    model = Invoice
-    permission_required = "change_challenge"
-    template_name = "challenges/invoice_list.html"
-    raise_exception = True
-    login_url = reverse_lazy("account_login")
-
-    def get_permission_object(self):
-        return self.request.challenge
-
-    def get_context_data(self, *args, **kwargs):
-        context = super().get_context_data(*args, **kwargs)
-
-        context["outstanding_invoices"] = any(
-            (
-                obj.payment_type != obj.PaymentTypeChoices.COMPLIMENTARY
-                and obj.payment_status == obj.PaymentStatusChoices.ISSUED
-            )
-            for obj in context["object_list"]
-        )
-
-        return context
-
-    def get_queryset(self):
-        queryset = super().get_queryset()
-        queryset = queryset.filter(challenge=self.request.challenge)
-        return queryset
